+++ src/main/kotlin/io/kjson/pointer/Extension.kt
@@ -25,6 +25,9 @@
 
 package io.kjson.pointer
 
+import kotlin.reflect.KClass
+import kotlin.reflect.typeOf
+
 import io.kjson.JSONArray
 import io.kjson.JSONObject
 import io.kjson.JSONValue
@@ -39,7 +42,7 @@ inline infix fun <reified T : JSONValue> JSONValue.ptr(pointer: JSONPointer): JS
  * Conditionally execute if [JSONObject] referenced by `this` [JSONRef] contains a member with the specified key and the
  * expected type.
  */
-inline fun <reified T : JSONValue> JSONRef<JSONObject>.ifPresent(name: String, block: JSONRef<T>.(T) -> Unit) {
+inline fun <reified T : JSONValue?> JSONRef<JSONObject>.ifPresent(name: String, block: JSONRef<T>.(T) -> Unit) {
     if (hasChild<T>(name))
         child<T>(name).let { it.block(it.node) }
 }
@@ -47,13 +50,97 @@ inline fun <reified T : JSONValue> JSONRef<JSONObject>.ifPresent(name: String, b
 /**
  * Iterate over the members of the [JSONObject] referenced by `this` [JSONRef].
  */
-inline fun <reified T : JSONValue> JSONRef<JSONObject>.forEachKey(block: JSONRef<T>.(String) -> Unit) {
+inline fun <reified T : JSONValue?> JSONRef<JSONObject>.forEachKey(block: JSONRef<T>.(String) -> Unit) {
     node.keys.forEach { child<T>(it).block(it) }
 }
 
 /**
  * Iterate over the members of the [JSONArray] referenced by `this` [JSONRef].
  */
-inline fun <reified T : JSONValue> JSONRef<JSONArray>.forEach(block: JSONRef<T>.(Int) -> Unit) {
+inline fun <reified T : JSONValue?> JSONRef<JSONArray>.forEach(block: JSONRef<T>.(Int) -> Unit) {
     node.indices.forEach { child<T>(it).block(it) }
+}
+
+/**
+ * Get the named child reference (strongly typed) from this [JSONObject] reference, using the implied child type.
+ */
+@Suppress("UNCHECKED_CAST")
+inline fun <reified Q : JSONValue?> JSONRef<JSONObject>.child(name: String): JSONRef<Q> {
+    val type = typeOf<Q>()
+    val childClass = type.classifier as KClass<JSONValue>
+    return child(childClass, name, type.isMarkedNullable) as JSONRef<Q>
+}
+
+/**
+ * Get the named child reference (strongly typed) from this [JSONObject] reference, using the supplied child class and
+ * nullability.
+ */
+fun <T : JSONValue> JSONRef<JSONObject>.child(
+    childClass: KClass<T>,
+    name: String,
+    nullable: Boolean = false,
+): JSONRef<T?> {
+    checkName(name)
+    return createTypedChildRef(childClass, nullable, node[name], name)
+}
+
+/**
+ * Get the named child reference (strongly typed) from this [JSONArray] reference, using the implied child type.
+ */
+@Suppress("UNCHECKED_CAST")
+inline fun <reified Q : JSONValue?> JSONRef<JSONArray>.child(index: Int): JSONRef<Q> {
+    val type = typeOf<Q>()
+    val childClass = type.classifier as KClass<JSONValue>
+    return child(childClass, index, type.isMarkedNullable) as JSONRef<Q>
+}
+
+/**
+ * Get the named child reference (strongly typed) from this [JSONArray] reference, using the supplied child class and
+ * nullability.
+ */
+fun <T : JSONValue> JSONRef<JSONArray>.child(
+    childClass: KClass<T>,
+    index: Int,
+    nullable: Boolean = false,
+): JSONRef<T?> {
+    checkIndex(index)
+    return createTypedChildRef(childClass, nullable, node[index], index.toString())
+}
+
+/**
+ * Get the named child reference (untyped) from this [JSONObject] reference.
+ */
+fun JSONRef<JSONObject>.untypedChild(name: String): JSONRef<JSONValue?> {
+    checkName(name)
+    return createChildRef(name, node[name])
+}
+
+/**
+ * Get the named child reference (untyped) from this [JSONArray] reference.
+ */
+fun JSONRef<JSONArray>.untypedChild(index: Int): JSONRef<JSONValue?> {
+    checkIndex(index)
+    return createChildRef(index.toString(), node[index])
+}
+
+/**
+ * Test whether this [JSONObject] reference has the named child.
+ */
+inline fun <reified T : JSONValue?> JSONRef<JSONObject>.hasChild(name: String): Boolean =
+        node.containsKey(name) && node[name] is T
+
+/**
+ * Test whether this [JSONArray] reference has a child at the given index.
+ */
+inline fun <reified T : JSONValue?> JSONRef<JSONArray>.hasChild(index: Int): Boolean =
+        index >= 0 && index < node.size && node[index] is T
+
+internal fun JSONRef<JSONObject>.checkName(name: String) {
+    if (!node.containsKey(name))
+        JSONPointer.pointerError("Node does not exist", "$pointer/$name")
+}
+
+internal fun JSONRef<JSONArray>.checkIndex(index: Int) {
+    if (index !in node.indices)
+        JSONPointer.pointerError("Index not valid", "$pointer/$index")
 }